Handle the arbitrary precision of integers in user roles.

+++ src/main/java/com/fastersheep/fastersheep/model/RoleEnum.java
@@ -1,11 +1,11 @@
 package com.fastersheep.fastersheep.model;
 
+import java.math.BigInteger;
+
 public enum RoleEnum {
     // FAQ:
     //
-    // This is a work-around for managing a few roles in RoleEnum file.
-    // In the real world, you do not need to cater for 
-    // more than 30 roles in a small organization.
+    // This is for managing an arbitrary number of roles in RoleEnum file.
     //
     //
     // Values here are unique prime numbers starting from 2.
@@ -16,55 +16,57 @@ public enum RoleEnum {
     // In an opinionated manner, using prime numbers
     // to map the permissions is better than 
     // database table joins because
-    // the process is simpler and easier to maintain.
+    // the process is done via BigInteger in this application
+    // and is easier to maintain.
     // There is no need to use @ManyToMany / @JoinColumn annotations
     // unless you need to store many roles and
     // you need to work with "table joins."
     //
     //
     // Example:
+    //         for (RoleEnum roleEnum : RoleEnum.values()) {
+    //                  if (roles.mod(roleEnum.getValue()) == 
+    //                  BigInteger.valueOf(0)) {
+    //                  rolesArr.add("ROLE_" + roleEnum.name());
+    //              }
+    //          }
     //
-    // 2*3 = 6 Then, 6 % 2 == 0 means DBADMIN
-    //               6 % 3 == 0 means ADMIN
-    //
-    // 7*11 = 77 Then, 77 % 7 == 0 means USER
-    //                 77 % 11 == 0 means GUEST
     //
     // There is also the file called RoleConfig
     // where the hierarchy can be modified / switched off.
     //
 
-    DBADMIN(2),
-    CEO(3),
-    CFO(5),
-    CTO(7),
-    ADMIN(11),
-    STAFF(13),
-    DEPARTMENTA(17),
-    DEPARTMENTB(19),
-    DEPARTMENTC(23),
-    DEPARTMENTD(29),
-    DEPARTMENTE(31),
-    DEPARTMENTF(37),
-    DEPARTMENTG(41),
-    DEPARTMENTH(43),
-    DEPARTMENTI(47),
-    DEPARTMENTJ(53),
-    DEPARTMENTK(59),
-    DEPARTMENTL(61),
-    DEPARTMENTM(67),
-    DEPARTMENTN(71),
-    DEPARTMENTO(73),
-    USER(79),
-    GUEST(83),
-    NOBODY(89);
+    DBADMIN(BigInteger.valueOf(2)),
+    CEO(BigInteger.valueOf(3)),
+    CFO(BigInteger.valueOf(5)),
+    CTO(BigInteger.valueOf(7)),
+    ADMIN(BigInteger.valueOf(11)),
+    STAFF(BigInteger.valueOf(13)),
+    DEPARTMENTA(BigInteger.valueOf(17)),
+    DEPARTMENTB(BigInteger.valueOf(19)),
+    DEPARTMENTC(BigInteger.valueOf(23)),
+    DEPARTMENTD(BigInteger.valueOf(29)),
+    DEPARTMENTE(BigInteger.valueOf(31)),
+    DEPARTMENTF(BigInteger.valueOf(37)),
+    DEPARTMENTG(BigInteger.valueOf(41)),
+    DEPARTMENTH(BigInteger.valueOf(43)),
+    DEPARTMENTI(BigInteger.valueOf(47)),
+    DEPARTMENTJ(BigInteger.valueOf(53)),
+    DEPARTMENTK(BigInteger.valueOf(59)),
+    DEPARTMENTL(BigInteger.valueOf(61)),
+    DEPARTMENTM(BigInteger.valueOf(67)),
+    DEPARTMENTN(BigInteger.valueOf(71)),
+    DEPARTMENTO(BigInteger.valueOf(73)),
+    USER(BigInteger.valueOf(79)),
+    GUEST(BigInteger.valueOf(83)),
+    NOBODY(BigInteger.valueOf(89));
 
-    public final long value;
+    public final BigInteger value;
 
-    private RoleEnum(long value) {
+    private RoleEnum(BigInteger value) {
         this.value = value;
     }
-    public long getValue() {
+    public BigInteger getValue() {
         return value;
     }
 }
